Reserve registers 0-2 for PC, EBP, ESP
EBP and ESP are read-only for the time being, since I don't see much
reason trying to sync changes to them. May just remove EBP and ESP
since I don't have anything to dereference the stack by the contents of
a register or something. Could add it, don't feel like it's useful yet.

This also adds the ability to refer to stack indices in registers
relative to EBP. All uses of StackIndex are now relative to either the
top of the stack or EBP. This reduces the range of argB to -1024 to
1023. argA is further restricted if accessing the stack (which is
already an unusual case for argA, so it may be removed) from -16 to 15,
so it's lost even more use.

Change-Id: Id512787a6cdf4387827541019db11e845d2c0768