Add OpGrow/Shrink
These are both fairly simple, but at least allow the option of
modifying stack size/capacity without write access to EBP/ESP.

OpShrink will effectively discard a chunk of the top of the stack, like
it probably suggests. OpGrow will only increase the capacity of the
stack, but not its actual size. If there's ever a reason to actual
alter the size in addition to capacity, I'll be amazed.

Change-Id: I123fa1e8d73891e21f6aa869fdf059eddcb6ae2a