Add arithmetic/bitwise shift tests
Also adds a slightly general-purpose way of testing different bits of
thread state following a run. Only applies to indexed bits of thread
state, however.

This doesn't cover exact type conversions (i.e., where the input is the
type it wants to begin with). It does ensure that vnum->vint works and
that the original type is preserved regardless of signed-ness.

Change-Id: Ied936bbecb30edd800e1cd8d0d60f3c17114a645