Consolidate rounding instructions
Rounding operands are now argAU for the rounding mode and argB for the
input. The output register is the same.

This chops 3 opcodes off the list and makes it a lot easier to simplify
things. In practice, rounding is also probably not necessary as an
instruction, but it does mean certain functions are free.

The only thing that might be necessary later is conversion
instructions, but I'm not sure those should be built into the VM. Same
thought around why the rounding instruction may not even be necessary:
rounding is a function, so what value is there in it being part of the
instruction set except to save function call costs? (And that might be
justifiable, but still, it's a little weird.)

Change-Id: If3695bbe895677e2efc74848da25c3212e6fdcd4