Change basic instruction format
Increases register count to 256. May go back to 127 to regain a couple
extra bits in instructions for other uses.

This also removes the OpShrink instruction, which is now implemented by
reducing ESP (i.e., add %esp %esp -5 will shrink the stack). EBP is
still read-only, but may make it possible to do alloca-like behavior
even if the stack isn't normal memory in Rusalka. The load instruction
will probably get a special format to allow loads into arbitrary stack
indices, as its argA is otherwise still too small, and we've already
lost a bit on argB and stack access through argA to the increased
register size.

OpLE, OpEQ, and OpLT are replaced with OpTest, which actually has
a disturbing range on its operands because there's no reason to include
a jump offset in the instruction.

Change-Id: If3d28fd6506c3210633f5e41701a69c88e88ff90