Mark 16 registers as frame-local
This means those 16 registers are preserved by allocating a stack frame
and restored after a call is complete. So, there's a small arena for
local variables to work in and be implicitly preserved as part of
a function call. If you need to preserve more than 16 registers, then
you'd need to a) push them onto the stack and b) make sure to pop them
later.

This is in line with existing Rusalka behavior, although a little
easier to fiddle with since the number of registers is just a small
number inside the existing 64 registers. It could be all 64, it could
be none.

Change-Id: Iade99a4e04134adfe843ec42a9c0cfabda5d7a43