Tweak push/pop instruction
Push/Pop now have their own instruction format, shared between the two.
There's an unused bit in pop as a result of this, but it seems like
a small price to pay.

This also changes the Example test to an actual test with subtests
(Go 1.7 is required by this) because mohae suggested it and it got kind
of annoying to figure out where in the list of prints/results two
matching rows were.

Change-Id: I6292479da5246434a2cb9d14265220282a50e3e7